feat: add inline snapshot testing framework (for lexer)
