+++ inline/tests/spec/snapshots/lexer/bold/bold-not-bold.snap.new
@@ -1,0 +1,23 @@
+---
+source: inline/tests/lexer/mod.rs
+assertion_line: 104
+info: "Test 'bold-not-bold' from: 'bold.yml'"
+---
+**Bold** not bold.
+**
+^^ -> Bold @ (1:1)->(1:3)
+  Bold
+  ^^^^ -> Plain @ (1:3)->(1:7)
+      **
+      ^^ -> Bold @ (1:7)->(1:9)
+         not bold.
+        ^^^^^^^^^^ -> Plain @ (1:9)->(1:19)
+                  ⏎
+                  ^ -> Newline @ (1:19)->(2:1)
+
+---
+With input:
+
+**Bold** not bold.
+
+
