+++ inline/tests/spec/snapshots/lexer/bold/escaped-bold.snap.new
@@ -1,0 +1,17 @@
+---
+source: inline/tests/lexer/mod.rs
+assertion_line: 104
+info: "Test 'escaped-bold' from: 'bold.yml'"
+---
+\*\*not bold\*\*
+**not bold**
+^^^^^^^^^^^^^^^^ -> Plain @ (1:1)->(1:17)
+                ⏎
+                ^ -> Newline @ (1:17)->(2:1)
+
+---
+With input:
+
+\*\*not bold\*\*
+
+
