+++ inline/tests/spec/snapshots/lexer/bold/not-bold.snap.new
@@ -1,0 +1,21 @@
+---
+source: inline/tests/lexer/mod.rs
+assertion_line: 104
+info: "Test 'not-bold' from: 'bold.yml'"
+---
+** not bold **
+**
+^^ -> Plain @ (1:1)->(1:3)
+   not bold 
+  ^^^^^^^^^^ -> Plain @ (1:3)->(1:13)
+            **
+            ^^ -> Plain @ (1:13)->(1:15)
+              ⏎
+              ^ -> Newline @ (1:15)->(2:1)
+
+---
+With input:
+
+** not bold **
+
+
