+++ inline/tests/spec/snapshots/lexer/bold/not-closed-bold.snap.new
@@ -1,0 +1,21 @@
+---
+source: inline/tests/lexer/mod.rs
+assertion_line: 104
+info: "Test 'not-closed-bold' from: 'bold.yml'"
+---
+**not bold **
+**
+^^ -> Plain @ (1:1)->(1:3)
+  not bold 
+  ^^^^^^^^^ -> Plain @ (1:3)->(1:12)
+           **
+           ^^ -> Plain @ (1:12)->(1:14)
+             ⏎
+             ^ -> Newline @ (1:14)->(2:1)
+
+---
+With input:
+
+**not bold **
+
+
