+++ inline/tests/spec/snapshots/lexer/bold/simple-bold.snap.new
@@ -1,0 +1,21 @@
+---
+source: inline/tests/lexer/mod.rs
+assertion_line: 104
+info: "Test 'simple-bold' from: 'bold.yml'"
+---
+**Bold**
+**
+^^ -> Bold @ (1:1)->(1:3)
+  Bold
+  ^^^^ -> Plain @ (1:3)->(1:7)
+      **
+      ^^ -> Bold @ (1:7)->(1:9)
+        ⏎
+        ^ -> Newline @ (1:9)->(2:1)
+
+---
+With input:
+
+**Bold**
+
+
