+++ inline/tests/spec/snapshots/lexer/plain/simple-plain.snap.new
@@ -1,0 +1,17 @@
+---
+source: inline/tests/lexer/mod.rs
+assertion_line: 53
+info: "Test 'simple-plain' from: 'plain.yml'"
+---
+Simple text.
+Simple text.
+^^^^^^^^^^^^ -> Plain @ (1:1)->(1:13)
+            ⏎
+            ^ -> Newline @ (1:13)->(2:1)
+
+---
+With input:
+
+Simple text.
+
+
